% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCovidData.R
\name{accumulateCovidData}
\alias{accumulateCovidData}
\title{Accumulate Tidy Coronavirus Data}
\usage{
accumulateCovidData()
}
\value{
Output is a dataframe with columns for date (Date), location (character),
  value_type (character; either "cases", "deaths", or "recoveries"), and value (int)
}
\description{
Imports data from public Johns Hopkins CSSEGISandData github repo on coronavirus
  confirmed cases, deaths, and recoveries by date and location, then collects all information
  in a tidy dataset
}
\examples{
importCovidData()

}
