% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSARSData.R
\name{cleanSARSData}
\alias{cleanSARSData}
\title{Loads and cleans SARS disease data}
\usage{
cleanSARSData()
}
\value{
Output is a dataframe with columns for date (Date), region (character),
  value_type (character; either "cases", "deaths", or "recovered"), and value (int)
}
\description{
Tidies and imports the SARS disease data to conform to the tidy format,
  does NOT include population data or latitude/longitude data
}
\examples{
cleanSARSData()

}
