% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZikaData.R
\name{cleanZikaData}
\alias{cleanZikaData}
\title{Clean Zika Data}
\usage{
cleanZikaData()
}
\value{
Output is a dataframe with columns for province, region, date, value, and value_type
}
\description{
Imports data from public github repo on Zika cases by date and location, 
  reformats the data into date/location/values type combinations, and cleans
  the values_type column to give proxies for confirmed cases for each row
}
\examples{
cleanZikaData()

}
