% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattDataAugmentation.R
\name{congregateDataDates}
\alias{congregateDataDates}
\title{Congregation of Data Dates}
\usage{
congregateDataDates(df)
}
\arguments{
\item{df}{A data frame for which the data values will be congregated to weekly dates.}
}
\value{
Output is a data frame of the same format as the passed-in df, but with congregated dates.
}
\description{
This function aligns and congregates data value counts on a weekly basis (taking the maximum if necessary)
so that the dates in the data are aligned for all data value types at all locations.
}
\details{
WARNING: This function manipulates the data in a way that makes the dates of observations 
less accurate, but in a way that is beneficial for plotting and understanding the cumulative
growth of the data.
}
\examples{
zika_raw <- importZikaData()
zika_2015 <- filterDiseaseData(zika_raw, last_date = "2015-12-31")
congregateDataDates(zika_2015)

}
