% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattDataAugmentation.R
\name{dayOfDiseaseColumn}
\alias{dayOfDiseaseColumn}
\title{Adding Day of Disease Column}
\usage{
dayOfDiseaseColumn(df, threshold = 100)
}
\arguments{
\item{df}{a data frame for which this day of disease column will be created.
This data frame must contain a column labeled as "date" which is of type "Date"}

\item{threshold}{the number of cases required to state the "start" of the disease.
Default is 100, which conforms with epidemiology convention.}
}
\value{
Output is the passed-in data frame with the addition of a day_of_disease column.
}
\description{
Adds a column of the data frame that gives the day of the disease, with day 1
being the first date seen in the passed-in data frame in which the threshold number
of cases was reached in the value column of the data frame.
}
\examples{
covid_data <- importCovidData()
dayOfDiseaseColumn(covid_data)

covid_us <- filterDiseaseData(covid_data, country = "US")
dayOfDiseaseColumn(covid_us)

zika_data <- filterDiseaseData(importZikaData())
dayOfDiseaseColumn(zika_data)

sars_data <- importSARSData()
dayOfDiseaseColumn(sars_data)

}
