% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterDiseaseData}
\alias{filterDiseaseData}
\title{Filter Data Frame}
\usage{
filterDiseaseData(data, first_date = NA, last_date = NA,
  country = c(), prov = c(), type = c("cases", "deaths",
  "recovered"), min_value = 0, max_value = Inf,
  include_suspected = FALSE)
}
\arguments{
\item{data}{A data frame to be filtered.}

\item{first_date}{The earliest date that is to be included in the returned data frame.}

\item{last_date}{The latest date that is to be included in the returned data frame.}

\item{country}{A vector of countries whose observations are to be included in the returned data frame.}

\item{prov}{A vector of provinces within the countries included in the regions vector
whose observations are to be included in the returned data frame.}

\item{type}{A vector of value types whose obbservations are to be included in the returned
data frame. The options are "cases", "deaths", and "recovered", and it defaults to include all
three if no input is given.}

\item{min_value}{The minimum value that is to be included in the returned data frame.}

\item{max_value}{The maximum value that is to be included in the returned data frame.}

\item{include_suspected}{Only applicable if disease = "zika". Boolean that determines whether suspected
cases should be included in cases. Default is FALSE.}
}
\value{
Output is the passed-in data frame filtered to the chosen specifications.
}
\description{
Allows for easy filtering of the data by:
Date - option to input first/last date
Region - Country/ies of interest
Province - Province(s) of interest within the country/ies of interest
Value Type - options are "cases", "deaths", and "recovered", defaults to include all three
Value - option to input minimum/maximum values
}
\examples{
filterDiseaseData(importCovidData(from_web = FALSE), country = c("Japan", "Italy", "US"))
filterDiseaseData(importSARSData(), type = c("cases", "deaths"), min_value = 10)
filterDiseaseData(importZikaData(), first_date = "2016-05-01", last_date = "2016-05-31")

}
