% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_global_data}
\alias{get_global_data}
\title{Query the global data online}
\usage{
get_global_data()
}
\value{
A 'global_summary' object
}
\description{
Query the global data online
}
\examples{
\dontrun{
x <- get_global_data()
}
}
