% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Query the latest data online}
\usage{
get_latest_data()
}
\value{
A 'nCov2019' object
}
\description{
Query global latest statistic for all contries
}
\author{
Guangchuang Yu
}
