% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCov2019.R
\name{get_nCov2019}
\alias{get_nCov2019}
\title{get_nCov2019}
\usage{
get_nCov2019(lang = "zh")
}
\arguments{
\item{lang}{one of 'zh' and 'en', for setting language of province and city names}
}
\value{
nCov2019 object
}
\description{
download statistical numbers of the wuhan 2019-nCov
}
\author{
Guangchuang Yu
}
