% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_therapeutics_data}
\alias{get_therapeutics_data}
\title{Query the therapeutics info online}
\usage{
get_therapeutics_data()
}
\value{
A 'vaccine_therapeutics' object
}
\description{
Query the therapeutics info online
}
\examples{
\dontrun{
x <- get_therapeutics_data()
}
}
