% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_vaccine_data}
\alias{get_vaccine_data}
\title{Query the vaccine info online}
\usage{
get_vaccine_data()
}
\value{
A 'vaccine_therapeutics' object
}
\description{
Query the vaccine info online
}
\examples{
\dontrun{
x <- get_vaccine_data()
}
}
