% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattDataAugmentation.R
\name{importCoordinateData}
\alias{importCoordinateData}
\title{Imports Coordinate Data}
\usage{
importCoordinateData()
}
\value{
Output is a dataframe with columns for country (character), 
latitude (double), and longitude (double).
}
\description{
Imports country latitude and longitude data from Google public dataset
}
\examples{
importCoordinateData()

}
