% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCovidData.R
\name{importCovidData}
\alias{importCovidData}
\title{Import Covid Data}
\usage{
importCovidData(from_web = T)
}
\arguments{
\item{from_web}{defaults to TRUE: whether to import from the web or from the package}
}
\value{
Output is a dataframe with columns for disease (covid), province (location specific), 
  region (location general), lat (latitude), long (longitude), date, value, value_type, pop_2018
}
\description{
Imports data from public github repo on Covid cases by date and location 
  (with latitude and longitude), reformats, cleans, and merges with population data
}
\examples{
importCovidData()               # defaults to from_web = T
importCovidData(from_web = F)   # returns historical subset of data

}
