% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSARSData.R
\name{importSARSData}
\alias{importSARSData}
\title{Import SARS data}
\usage{
importSARSData(from_web = F)
}
\arguments{
\item{from_web}{defaults to FALSE: whether to import from the web or from the package}
}
\value{
Output is a dataframe with columns for disease (sars),
region (country), value, value_type, pop_2016, lat (latitude), long (longitude)
}
\description{
Imports data from public github repo on SARS cases by date and location,
  reformats and cleans the data, and merges with population and latitude-longitude data
}
\examples{
importSARSData()

}
