% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZikaData.R
\name{importZikaData}
\alias{importZikaData}
\title{Import Zika Data}
\usage{
importZikaData(from_web = F)
}
\arguments{
\item{from_web}{defaults to FALSE: whether to import from the web or from the package}
}
\value{
Output is a dataframe with columns for disease (zika), province (location specific), 
region (location general), value, value_type, pop_2016, lat (latitude), long (longitude)
}
\description{
Imports data from public github repo on Zika cases by date and location,
  reformats, cleans, and merges with population and latitude-longitude data
}
\examples{
importZikaData()   # from_web defaults to FALSE
importZikaData(from_web = T) 

}
