% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCov2019.R
\name{load_nCov2019}
\alias{load_nCov2019}
\title{load_nCov2019}
\usage{
load_nCov2019(lang = "auto", source = "github")
}
\arguments{
\item{lang}{one of 'zh' and 'en', for setting language of province and city names.
If lang = "auto" (default), it will be set based on Sys.getlocale("LC_CTYPE")}

\item{source}{historical data source, one of 'github' or 'dxy'}
}
\value{
nCov2019History object
}
\description{
load historical data of COVID-2019
}
\author{
Guangchuang Yu
}
