% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattDataViz.R
\name{mapPlotAnimate}
\alias{mapPlotAnimate}
\title{Animated Map Plot}
\usage{
mapPlotAnimate(data, first_date = NA, last_date = NA,
  selected_value_type = NA, color = "red", alpha = 0.5, fps = 10)
}
\arguments{
\item{data}{The data frame that contains the data to be plotted on the world map.
This data frame must contain the following columns: "long" (type "numeric"), 
"lat" (type "numeric"), and "value" (type "numeric").}

\item{first_date}{This is the first date shown in the animation.}

\item{last_date}{This is the last date shown in the animation.}

\item{selected_value_type}{If the data frame contains a column "value_type" with multiple dates,
this parameter must specify a specific data value type to be plotted. Options are "cases", "deaths", 
and "recovered".}

\item{color}{Color of points on world map. Default is red.}

\item{alpha}{Transparency level of points on world map. Default is 0.5.}

\item{fps}{Number of frames displayed per second for the animation. Default is 10.}
}
\value{
Output is an animated ggplot object which maps the chosen value on a world map using
latitude and longitude data over the range of dates specified.
}
\description{
Generates an animated global map plot of a certain value of data.
}
\details{
Note: Expect inconsequential warnings when running this function that do not deter the 
construction or alter the behavior of the animation plot in any way.
}
\examples{
covid_data <- importCovidData()
sars_data <- importSARSData()
zika_data <- filterDiseaseData(importZikaData(), include_suspected = FALSE)
mapPlotAnimate(covid_data, fps = 6)
mapPlotAnimate(sars_data, selected_value_type = "recovered")
mapPlotAnimate(zika_data, color = "dark green", alpha = 0.7)

}
