% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Shiny app}
\usage{
open_dashboard(lang = "auto", remote = FALSE, install_it = FALSE)
}
\arguments{
\item{lang}{Your languate, it should be one of "auto", "zh" or "en"}

\item{remote}{If TRUE, open the online version}

\item{install_it}{If TRUE, the function will install packages needed}
}
\description{
a shiny app
}
