% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nCov2019History}
\alias{plot.nCov2019History}
\title{plot.nCov2019History}
\usage{
\method{plot}{nCov2019History}(
  x,
  region = "Global",
  continuous_scale = TRUE,
  palette = "Reds",
  date = NULL,
  from = NULL,
  to = NULL,
  width = 600,
  height = 600,
  filename = "nCov2019.gif",
  fps = 2,
  type,
  ...
)
}
\arguments{
\item{from}{start date to plot}

\item{to}{end date to plot. Both from and to should be specify, otherwise they will be ignored.
If both from and to are specify, an animation will be created.}

\item{width}{width of the plot, only works for animation}

\item{height}{height of the plot, only works for animation}

\item{fps}{fps of the animation, only works for animation}
}
\description{
plot.nCov2019History
}
