% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZikaData.R
\name{reformatZikaData}
\alias{reformatZikaData}
\title{Reformat Zika Data}
\usage{
reformatZikaData()
}
\value{
Output is a dataframe with columns for province, region, date, value, and value_type
}
\description{
Imports data from public github repo on Zika cases by date and location, 
  then reformats the data into date/location/values type combinations
}
\examples{
reformatZikaData()

}
