% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sars_data}
\alias{sars_data}
\title{SARS Data}
\format{A tibble with 31551 rows and 8 columns
\describe{
  \item{disease} "sars" for all rows
  \item{region} a general location identifier, usually the country
  \item{date} the date of the observation
  \item{value} an integer representing the quantity of value_type for that observation
  \item{value_type} either "cases", "deaths", or "recovered"
  \item{pop_2003} region population data for 2003
  \item{lat} the latitude of the region
  \item{long} the longitude of the region
}}
\source{
\url{https://raw.githubusercontent.com/mcolon21/ncov2019data/master/sars_data.csv}
}
\description{
A dataset of SARS virus data with rows being location/date/value_type combinations, 
  columns being identifying information and the associated value
}
