% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCovidData.R
\name{scrapeCovidData}
\alias{scrapeCovidData}
\title{Downloads Time-Series Coronavirus Data}
\usage{
scrapeCovidData(type = "cases")
}
\arguments{
\item{type}{a string representing the type of coronavirus data desired:
either "cases" for confirmed cases (default), "deaths" for deaths,
or "recovered" for recoveries}
}
\value{
Output is a dataframe with rows representing locations,
  columns representing dates, values representing the number of people
}
\description{
Imports data from public Johns Hopkins CSSEGISandData github repo on
  day-by-day coronavirus confirmed cases, deaths, and recoveries by location
}
\examples{
importCovidData()
importCovidData("cases")
importCovidData("deaths")
importCovidData("recovered") 

}
