% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSARSData.R
\name{scrapeSARSData}
\alias{scrapeSARSData}
\title{Scrapes SARS disease data}
\usage{
scrapeSARSData()
}
\value{
Output is a raw dataframe in its untidy format, directly from
 the Github repository
}
\description{
Scrapes the Kaggle SARS disease data from a public Github repository.
}
\examples{
scrapeSARSData()

}
