% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZikaData.R
\name{scrapeZikaData}
\alias{scrapeZikaData}
\title{Downloads Zika virus data}
\usage{
scrapeZikaData()
}
\value{
Output is a dataframe with rows for date-location-case_type pairs
}
\description{
Imports the Kaggle Zika virus data from a public Github repository.
}
\examples{
scrapeZikaData()

}
