% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCovidData.R
\name{tidyCovidData}
\alias{tidyCovidData}
\title{Tidies up scraped coronavirus data}
\usage{
tidyCovidData(covid_df, covid_df_name)
}
\arguments{
\item{covid_df}{a time-series dataframe of the type scraped from 
JHUCSSEGISandData github repo (columns for dates, rows for locations)}

\item{covid_df_name}{a character string representing the type of coronavirus data:
ie either "cases", "deaths", or "recovered"}
}
\value{
Output is a dataframe with columns for province, region, lat, long,
  date, value, and value_type
}
\description{
Pivots the data frame to make one row per location/date combo
}
\examples{
tidyCovidData(scrapeCovidData())
tidyCovidData(scrapeCovidData("cases"))
tidyCovidData(scrapeCovidData("deaths"))

}
