% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{trans_city}
\alias{trans_city}
\title{trans_city}
\usage{
trans_city(city, lang = "zh")
}
\arguments{
\item{city}{city name}

\item{lang}{language of city name. if 'zh', translate to 'en' and if 'en', translate to 'zh'}
}
\value{
translated city name
}
\description{
translate city from Chinese to English and vice versa
}
\author{
Patrick Tung and Guangchuang Yu
}
