% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{trans_province}
\alias{trans_province}
\title{trans_province}
\usage{
trans_province(province, lang = "zh")
}
\arguments{
\item{province}{province name}

\item{lang}{language of province name. if 'zh', translate to 'en' and if 'en', translate to 'zh'}
}
\value{
translated province name
}
\description{
translate province from Chinese to English and vice versa
}
\author{
Patrick Tung and Guangchuang Yu
}
