% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattDataAugmentation.R
\name{congregateDataDates}
\alias{congregateDataDates}
\title{Congregation of Data Dates}
\usage{
congregateDataDates(df)
}
\arguments{
\item{df}{a data frame for which the data values will be congregated to weekly dates.}
}
\value{
Output is a data frame of the same format as the passed-in df, but with congregated dates.
}
\description{
Because the Zika data was tracked on different dates in different locations, this function
congregates the value counts over a certain period (taking the maximum if necessary) so that
the dates in the data are aligned for all locations. This function is only to be used on Zika data.
}
\details{
WARNING: This function manipulates the data in a way that makes the dates less accurate,
but is beneficial for plotting.
}
\examples{
congregateDataDates(importZikaData())
congregateDataDates(filterDiseaseData(importZikaData(), include_suspected = FALSE))

}
