% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(data)
}
\arguments{
\item{data}{users' own data, it should contain these 6 column: "country","province","date","cases","deaths","recovered".}
}
\value{
a 'nCov2019History' object
}
\description{
Convert users' own data into class of nCov2019History data. Then it could be used in nCov2019.
}
