% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_pop_data_raw}
\alias{country_pop_data_raw}
\title{Population Data for Countries}
\format{A tibble with 265 rows and 5 columns
\describe{
  \item{country_name}
  \item{country_code}
  \item{pop_2003}
  \item{pop_2016}
  \item{pop_2018}
}}
\source{
\url{https://data.worldbank.org/indicator/SP.POP.TOTL}
}
\description{
A dataset with rows for countries (or groups of countries), columns for years in 
  which the diseases occurred (2003 for SARS, 2016 for Zika, 
  2018 for Covid (most recent year of data)), values are population
}
