% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_nCov2019}
\alias{get_nCov2019}
\title{Query the latest data}
\usage{
get_nCov2019()
}
\value{
The latest statistic data
}
\description{
Query the latest data online;deprecated, 
use get_latest_data() in the further.
}
\examples{
\dontrun{
x <- get_nCov2019()
}
}
