% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{load_nCov2019}
\alias{load_nCov2019}
\title{Query the latest data}
\usage{
load_nCov2019()
}
\value{
The historical statistic data
}
\description{
Query the historical data online;deprecated, 
use load_nCov2019() in the further.
}
\examples{
\dontrun{
x <- load_nCov2019()
}
}
