% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Shiny app}
\usage{
open_dashboard(lang = "auto", remote = FALSE)
}
\arguments{
\item{lang}{Your language}

\item{remote}{If TRUE, open the online version}
}
\description{
a shiny app
}
