% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silasDataViz.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Create Time-Series Plots}
\usage{
plotTimeSeries(data, plot_what = "cases", group = "all",
  x_axis = "date")
}
\arguments{
\item{data}{disease data to plot (usually the output of filterDiseaseData)}

\item{plot_what}{options are: "cases", "recovered", "deaths" (plots totals by date),
"log_cases", "log_recovered", "log_deaths" (plots log of totals by date),
"x_per_y" where x and y are either "cases", "recovered", "deaths", or "pop" (plots the ratio by date),
"new_cases", "new_recovered", "new_deaths" (plots change by date) or 
"growth_factor" (plots new cases divided by new cases on previous date for each date),
if plot_what is not specified, plots cases, recovered, and deaths by day}

\item{group}{options are: "province", "region", or "all" ("all" is the default).
"province" creates one plot layer for each province in the data,
"region" creates one plot layer for each region in the data,
"all" creates one plot layer total}

\item{x_axis}{options are "date" ("date" is the default) or "day_of_disease".
"date" makes the x-axis the calendar date,
"day_of_disease" makes the x-axis the days since the first case in the group
(where group is specified in the group argument)}
}
\description{
Visualize disease statistics on a day-by-day basis
}
\examples{
covid_data <- importCovidData()
us_covid_data <- filterDiseaseData(covid_data, country = "US")
plotTimeSeries(us_covid_data)
plotTimeSeries(us_covid_data, plot_what = "cases")
plotTimeSeries(us_covid_data, plot_what = "log_cases")
plotTimeSeries(us_covid_data, plot_what = "deaths_per_cases")
plotTimeSeries(us_covid_data, plot_what = "recovered_per_cases")
plotTimeSeries(us_covid_data, plot_what = "cases_per_pop")
plotTimeSeries(us_covid_data, plot_what = "new_cases")
plotTimeSeries(us_covid_data, plot_what = "new_deaths")
plotTimeSeries(us_covid_data, plot_what = "growth_factor")

}
