% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCovidData.R
\name{scrapeCovidData}
\alias{scrapeCovidData}
\title{Scrapes Time-Series Coronavirus Data}
\usage{
scrapeCovidData(type = "cases")
}
\arguments{
\item{type}{a string representing the type of coronavirus data desired:
either "cases" for confirmed cases (default), "deaths" for deaths,
or "recovered" for recoveries}
}
\value{
Output is a dataframe with rows representing locations,
  columns representing dates, values representing the number of cases/deaths/recovered
  for that location/date combination
}
\description{
Imports data from public Johns Hopkins CSSEGISandData github repo on
  day-by-day coronavirus confirmed cases, deaths, and recoveries by location
}
\examples{
scrapeCovidData()    # defaults to scrape "cases"
scrapeCovidData("cases")
scrapeCovidData("deaths")
scrapeCovidData("recovered") 

}
